requirements.txt
openai
whisper
moviepy
python-dotenv
pysrt
requests
google-genai 
streamlit
